requirements.txt
django==5.0.9
django-environ==0.11.2
django-meta==2.4.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
plotly==5.22.0
kaleido==0.2.1
pybeerxml==2.1.3
translitcodec==0.7.0
lxml==5.2.2
gunicorn==22.0.0
whitenoise==6.7.0
requests==2.32.3
